localhost -> 0.0.0.0 Additional dht tests
