+++ p2p/dht/table/tests/table_test.go
@@ -3,52 +3,12 @@ package tests
 import (
 	"github.com/spacemeshos/go-spacemesh/p2p"
 	"github.com/spacemeshos/go-spacemesh/p2p/dht/table"
-	"github.com/spacemeshos/go-spacemesh/p2p/node"
 	"math/rand"
 	"runtime"
 	"testing"
 	"time"
 )
 
-// Test basic features of the bucket struct
-func TestBucket(t *testing.T) {
-	const n = 100
-
-	local := p2p.GenerateRandomNodeData()
-	localId := local.DhtId()
-
-	b := table.NewBucket()
-	nodes := p2p.GenerateRandomNodesData(n)
-	for i := 0; i < n; i++ {
-		b.PushFront(nodes[i])
-	}
-
-	i := rand.Intn(len(nodes))
-	if !b.Has(nodes[i]) {
-		t.Errorf("Failed to find peer: %v", nodes[i])
-	}
-
-	spl := b.Split(0, localId)
-	llist := b.List()
-
-	for e := llist.Front(); e != nil; e = e.Next() {
-		id := e.Value.(node.RemoteNodeData).DhtId()
-		cpl := id.CommonPrefixLen(localId)
-		if cpl > 0 {
-			t.Fatalf("Split failed. found id with cpl > 0 in 0 bucket")
-		}
-	}
-
-	rlist := spl.List()
-
-	for e := rlist.Front(); e != nil; e = e.Next() {
-		id := e.Value.(node.RemoteNodeData).DhtId()
-		cpl := id.CommonPrefixLen(localId)
-		if cpl == 0 {
-			t.Fatalf("Split failed. found id with cpl == 0 in non 0 bucket")
-		}
-	}
-}
 
 func TestTableCallbacks(t *testing.T) {
 	const n = 100
@@ -105,7 +65,6 @@ Loop1:
 	}
 }
 
-// Right now, this just makes sure that it doesnt hang or crash
 func TestTableUpdate(t *testing.T) {
 
 	const n = 100
@@ -129,13 +88,13 @@ func TestTableUpdate(t *testing.T) {
 		// create callback to receive result
 		callback := make(table.PeersOpChannel, 2)
 
-		// find nearest peers
+		// find nearest peers to new node
 		rt.NearestPeers(table.NearestPeersReq{n.DhtId(), 5, callback})
 
 		select {
 		case c := <-callback:
-			if len(c.Peers) == 0 {
-				t.Fatalf("Failed to find node near %s.", n.DhtId())
+			if len(c.Peers) != 5 {
+				t.Fatalf("Expected to find 5 close nodes to %s.", n.DhtId())
 			}
 		case <-time.After(time.Second * 5):
 			t.Fatalf("Failed to get expected update callbacks on time")
@@ -162,8 +121,7 @@ func TestTableFind(t *testing.T) {
 
 		n := nodes[i]
 
-		//t.Logf("Searching for peer: %s...", hex.EncodeToString(node.DhtId()))
-
+		// try to find nearest peer to n - it should be n
 		callback := make(table.PeerOpChannel, 2)
 		rt.NearestPeer(table.PeerByIdRequest{n.DhtId(), callback})
