+++ p2p/net/network_test.go
@@ -15,7 +15,7 @@ func TestReadWrite(t *testing.T) {
 	msg := []byte("hello world")
 	msgId := crypto.UUID()
 	port := crypto.GetRandomUInt32(1000) + 10000
-	address := fmt.Sprintf("localhost:%d", port)
+	address := fmt.Sprintf("0.0.0.0:%d", port)
 	done := make(chan bool, 1)
 
 	n, err := NewNet(address, nodeconfig.ConfigValues)
